Fixes issue with error

+++ crates/cibse_guide_e/src/chapter_7/equation_7_6/openfire_runner.rs
@@ -73,7 +73,7 @@ impl MethodRunner for Chapter7Equation6Runner {
             .units("persons/s")
             .build();
 
-        let w = ParamBuilder::integer(&SYMBOLS.w)
+        let w = ParamBuilder::float(&SYMBOLS.w)
             .name("Width of the opening or corridor")
             .units("m")
             .min_exclusive(0.0)
